+++ aws/scripts/install_java8.sh
@@ -1,43 +1,51 @@
 #!/bin/bash
 
+if [[ -f /tmp/params.sh ]]; then 
+	source /tmp/params.sh
+fi
+
+if [[ -z ${JAVA_VERSION} || -z ${JAVA_HOME} || -z ${JAVA_DOWNLOAD_URL} || -z ${JCE_DOWNLOAD_URL} ]]; then 
+	echo ' ${JAVA_VERSION}, ${JAVA_HOME}, ${JAVA_DOWNLOAD_URL}, ${JCE_DOWNLOAD_URL} must be defined'
+	exit 1
+fi
+
+#Print params before starting
+echo "JAVA_VERSION: ${JAVA_VERSION}"
+echo "JAVA_HOME: ${JAVA_HOME}"
+echo "JAVA_DOWNLOAD_URL: ${JAVA_DOWNLOAD_URL}"
+echo "JCE_DOWNLOAD_URL: ${JCE_DOWNLOAD_URL}"
+
+
 #remove any openjdk
 yum remove --assumeyes *openjdk*
 
-export JAVA_HOME='/usr/java/jdk1.8.0_121'
+echo "==> Installing Java ${JAVA_VERSION} to ${JAVA_HOME}"
 
 if [[ -d ${JAVA_HOME} ]]; then
-	echo "jdk1.8.0_121 already installed. Skipping."
+	echo "${JAVA_VERSION} is already installed. ${JAVA_HOME} already exists. Skipping..."
 else
-	echo "Installing Java 8... "
-
-	wget --no-cookies --no-check-certificate \
-	        --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com%2F; oraclelicense=accept-securebackup-cookie" \
-	         "http://download.oracle.com/otn-pub/java/jdk/8u121-b13/e9e7ea248e2c4826b92b3f075a80e441/jdk-8u121-linux-x64.rpm" \
-	         -O /tmp/jdk-8u121-linux-x64.rpm
-
-	rpm -ivh /tmp/jdk-8u121-linux-x64.rpm
-
-
-	rm -rf /tmp/jdk-8u121-linux-x64.rpm
+	echo "Installing ${JAVA_VERSION}... "
+	JAVA_RPM=${JAVA_DOWNLOAD_URL##*/}
+	echo "Downloading ${JAVA_VERSION} from ${JAVA_DOWNLOAD_URL}"
+    wget -O /tmp/${JAVA_RPM} --no-cookies --no-check-certificate --header "${ORACLE_COOKIE}" ${JAVA_DOWNLOAD_URL}
+		    
+	
+	rpm -ivh /tmp/${JAVA_RPM}
+	rm -rf /tmp/${JAVA_RPM}
 	echo "Installing Java 8... Done"
 
-	echo "Installing Java 8 JCE... "
+	echo "Installing Java JCE 8 ... "
+	JCE_FILE=${JCE_DOWNLOAD_URL##*/}	
+			
 	###  Install Java Unlimited Strength Encryption Policy Files for Java 8  ###
-	wget -O /tmp/jce_policy-8.zip --no-cookies --no-check-certificate \
-	      --header "Cookie: gpw_e24=http%3A%2F%2Fwww.oracle.com%2F; oraclelicense=accept-securebackup-cookie" \
-	      "http://download.oracle.com/otn-pub/java/jce/8/jce_policy-8.zip"
-	unzip -o /tmp/jce_policy-8.zip -d /tmp
-	rm -f ${JAVA_HOME}/jre/lib/security/local_policy.jar
-	rm -f ${JAVA_HOME}/jre/lib/security/US_export_policy.jar
-	mv /tmp/UnlimitedJCEPolicyJDK8/local_policy.jar ${JAVA_HOME}/jre/lib/security/local_policy.jar
-	mv /tmp/UnlimitedJCEPolicyJDK8/US_export_policy.jar ${JAVA_HOME}/jre/lib/security/US_export_policy.jar
-
+    wget -O /tmp/${JCE_FILE} --no-cookies --no-check-certificate --header "${ORACLE_COOKIE}" ${JCE_DOWNLOAD_URL} 
+		
+			
+	unzip -jo /tmp/${JCE_FILE} -d ${JAVA_HOME}/jre/lib/security/	
 	rm -rf /tmp/jce_policy-8.zip
-	rm -rf /tmp/UnlimitedJCEPolicyJDK8
 	echo "Installing Java 8 JCE... Done"
 fi
 java -version
-
